refactor models to separate file
